+++ test/runtests.jl
@@ -4,14 +4,18 @@ using Unitless
 using Unitful
 using Test
 
-struct MyNumber{T<:Number} <: Number
+# Minimal implementation of a custom numeric type.
+struct MyNumber{T<:Unitless.BareNumber} <: Number
     val::T
 end
-
 Base.zero(::Type{MyNumber{T}}) where {T} = MyNumber{T}(zero(T))
 Base.oneunit(::Type{MyNumber{T}}) where {T} = MyNumber{T}(one(T))
 Base.one(::Type{MyNumber{T}}) where {T} = one(T)
-Unitless.unitless(x::MyNumber) = getfield(x,:val)
+Base.real(x::MyNumber{<:Real}) = x
+Base.real(x::MyNumber{<:Complex}) = MyNumber(real(x.val))
+Base.real(::Type{MyNumber{T}}) where {T<:Real} = MyNumber{T}
+Base.real(::Type{MyNumber{Complex{T}}}) where {T<:Real} = MyNumber{T}
+Unitless.unitless(x::MyNumber) = x.val
 
 @testset "Basic types" begin
     # bare_type with no argument
@@ -53,6 +57,50 @@ Unitless.unitless(x::MyNumber) = getfield(x,:val)
     @test convert_bare_type(Int, 2.0) === 2
     @test convert_bare_type(Float32, 2.0) === 2.0f0
     @test convert_bare_type(MyNumber{Int16}, 12.0) === Int16(12)
+    @test_throws ErrorException convert_bare_type(Int, "oups!")
+
+    # real_type with no argument
+    @test real_type() === Real
+
+    # real_type for values
+    @test real_type(1.0) === Float64
+    @test real_type(Float32) === Float32
+    @test real_type(Complex(2,3)) === Int
+    @test real_type(NaN) === typeof(NaN)
+    @test real_type(π) === typeof(π)
+    @test real_type(3//4) === typeof(3//4)
+    @test_throws ErrorException real_type("hello")
+
+    # real_type for types
+    @test real_type(Real) === Real
+    @test real_type(Integer) === Integer
+    @test real_type(Float32) === Float32
+    @test real_type(BigFloat) === BigFloat
+    @test real_type(Complex{Int}) === Int
+    @test real_type(typeof(π)) === typeof(π)
+    @test real_type(typeof(3//4)) === typeof(3//4)
+    @test_throws ErrorException real_type(AbstractString)
+
+    # real_type with multiple arguments
+    @test real_type(1, 0f0) === Float32
+    @test real_type(Int, pi) === promote_type(Int, typeof(pi))
+    @test real_type(4, pi, 1.0) === promote_type(Int, typeof(pi), Float64)
+    @test real_type(Int, Int8, Float32) === promote_type(Int, Int8, Float32)
+    @test real_type(Int, Int8, Float32) === promote_type(Int, Int8, Float32)
+    @test real_type(Int, Int8, Complex{Int16}, Float32) === promote_type(Int, Int8, Int16, Float32)
+
+    # default implementation
+    @test real_type(MyNumber(1.2f0)) === Float32
+    @test real_type(MyNumber{Complex{Int16}}) === Int16
+
+    # convert_real_type
+    @test convert_real_type(Int, -1) === -1
+    @test convert_real_type(Int, 2.0) === 2
+    @test convert_real_type(Complex{Float32}, 2.0) === 2.0f0
+    @test convert_real_type(Complex{Int16}, 2 + 3im) === Complex{Int16}(2, 3)
+    @test convert_real_type(Float32, 2.0 - 1.0im) === Complex{Float32}(2, -1)
+    @test convert_real_type(MyNumber{Int16}, 12.0) === Int16(12)
+    @test_throws ErrorException convert_real_type(Int, "oups!")
 
     # unitless
     @test unitless(Real) === bare_type(Real)
@@ -86,11 +134,31 @@ end
     # bare_type with multiple arguments
     @test bare_type(u"2.0m/s", u"35GHz") === Float64
     @test bare_type(1, u"2.0f0m/s", u"35GHz") === Float32
+    @test bare_type(1, u"2.0f0m/s", u"35GHz", Complex{Int8}(11)) === Complex{Float32}
 
     # convert_bare_type
     @test convert_bare_type(Float64, u"2.0m/s") === u"2.0m/s"
     @test convert_bare_type(Int, u"2.0m/s") === u"2m/s"
     @test convert_bare_type(Float32, u"35GHz") === u"35.0f0GHz"
+
+
+    # real_type for values
+    @test real_type(u"2.0m/s") === Float64
+    @test real_type(u"35GHz") === Int
+
+    # real_type for types
+    @test real_type(typeof(u"2.0m/s")) === Float64
+    @test real_type(typeof(u"35GHz")) === Int
+
+    # real_type with multiple arguments
+    @test real_type(u"2.0m/s", u"35GHz") === Float64
+    @test real_type(1, u"2.0f0m/s", u"35GHz") === Float32
+    @test real_type(1, u"2.0f0m/s", u"35GHz", Complex{Int8}(11)) === Float32
+
+    # convert_real_type
+    @test convert_real_type(Float64, u"2.0m/s") === u"2.0m/s"
+    @test convert_real_type(Int, u"2.0m/s") === u"2m/s"
+    @test convert_real_type(Float32, u"35GHz") === u"35.0f0GHz"
 
     # unitless
     @test unitless(u"17GHz") === 17